My code review (and edits) of fast probing. Yang, can you please do a code review of my code review/edits?
I removed the pingQ variable and pingAll() method as they are not currently being used and confused me for about 15 minutes until I realized this.

I combined the logic for setting isReachable etc. into the run() loop of the ProbeTask, removing the timeout that was previously in resetTimeoutAtNode() -- the latter didn't do quite what we wanted and had synchronization issues with regards to the ProbeTask.

Finally, *** I don't see anywhere where we probe the coordinator. *** This was previously in pingAll (which was never used). Do we need this? Will the coordinator take care of probing the members?

These changes should now *almost* be enough to properly take care of node failures, and it would be worth testing to check. I think the remaining necessary change is to change the amount of time (in getAllRendezvousServers) we wait to confirm a node failure from 1 routing iteration (r) to r+p+\epsilon, but perhaps this will turn out to be minor.





git-svn-id: https://moo.cmcl.cs.cmu.edu/svn/scaleron/src/java_sim@704 0bd769cf-3f23-0410-95ed-bb08bbc236b7